jjp-132: rename authSlice to registerSlice, authService -> registerService
