+++ app/frontend/src/redux/store.ts
@@ -1,13 +1,13 @@
 import { configureStore } from "@reduxjs/toolkit";
 import usersReducer from "./slices/usersSlice";
 import eventReducer from "./slices/eventSlice";
-import authReducer from "./slices/authSlice";
+import registerReducer from "./slices/registerSlice";
 
 export const store = configureStore({
   reducer: {
     users: usersReducer,
     event: eventReducer,
-    auth: authReducer
+    registerData: registerReducer,
   },
 });
